+++ demo/index.js
@@ -15,7 +15,10 @@ new Vue({  // eslint-disable-line
       value6: '',
       value7: '',
       value8: '',
-      value9: ''
+      value9: '',
+      value10: new Date(),
+      value11: new Date(),
+      value12: ''
     }
   },
   methods: {
@@ -44,7 +47,10 @@ new Vue({  // eslint-disable-line
   render (h) {
     const example1 = {
       'base': '<date-picker v-model="value1" lang="en" :not-before="new Date()"></date-picker>',
-      'range': '<date-picker v-model="value2" range ></date-picker>'
+      'range': '<date-picker v-model="value2" range ></date-picker>',
+      'month': '<date-picker v-model="value10" lang="en" type="month" format="YYYY-MM"></date-picker>',
+      'year': '<date-picker v-model="value11" lang="en" type="year" format="YYYY"></date-picker>',
+      'time': '<date-picker v-model="value12" lang="en" type="time" format="HH:mm:ss" placeholder="Select Time"></date-picker>'
     }
     const example2 = {
       'datetime': `
@@ -52,7 +58,7 @@ new Vue({  // eslint-disable-line
           v-model="value3"
           lang="en"
           type="datetime"
-          format="YYYY-MM-DD HH:mm:ss"></date-picker>`,
+          format="[on] MM-DD-YYYY [at] HH:mm"></date-picker>`,
       'datetime with time-picker-options': `
         <date-picker
           v-model="value4"