+++ src/calendar.vue
@@ -18,22 +18,26 @@
         class="mx-icon-next-month"
         @click="handleIconMonth(1)">&rsaquo;</a>
       <a
-        v-show="panel !== 'TIME'"
+        v-show="panel === 'DATE'"
         class="mx-current-month"
         @click="handleBtnMonth">{{months[calendarMonth]}}</a>
       <a
-        v-show="panel !== 'TIME'"
+        v-show="panel === 'DATE' || panel === 'MONTH'"
         class="mx-current-year"
         @click="handleBtnYear">{{calendarYear}}</a>
       <a
+        v-show="panel === 'YEAR'"
+        class="mx-current-year">{{yearHeader}}</a>
+      <a
         v-show="panel === 'TIME'"
         class="mx-time-header"
-        @click="showPanelDate">{{timeHeader}}</a>
+        @click="handleTimeHeader">{{timeHeader}}</a>
     </div>
     <div class="mx-calendar-content">
       <panel-date
         v-show="panel === 'DATE'"
         :value="value"
+        :date-format="dateFormat"
         :calendar-month="calendarMonth"
         :calendar-year="calendarYear"
         :start-at="startAt"
@@ -44,11 +48,13 @@
       <panel-year
         v-show="panel === 'YEAR'"
         :value="value"
+        :disabled-year="isDisabledYear"
         :first-year="firstYear"
         @select="selectYear" />
       <panel-month
         v-show="panel === 'MONTH'"
         :value="value"
+        :disabled-month="isDisabledMonth"
         :calendar-year="calendarYear"
         @select="selectMonth" />
       <panel-time
@@ -57,13 +63,15 @@
         :time-picker-options="timePickerOptions"
         :value="value"
         :disabled-time="isDisabledTime"
-        @select="selectTime" />
+        :time-type="timeType"
+        @select="selectTime"
+        @pick="pickTime" />
     </div>
   </div>
 </template>
 
 <script>
-import { isValidDate, isDateObejct } from '@/utils/index'
+import { isValidDate, isDateObejct, formatDate } from '@/utils/index'
 import locale from '@/mixins/locale'
 import scrollIntoView from '@/utils/scroll-into-view'
 import PanelDate from '@/panel/date'
@@ -88,12 +96,15 @@ export default {
       type: Boolean,
       default: false
     },
-
-    // below user set
     type: {
       type: String,
-      default: 'date' // ['date', 'datetime']
+      default: 'date' // ['date', 'datetime'] zendy added 'month', 'year', mxie added "time"
     },
+    dateFormat: {
+      type: String,
+      default: 'YYYY-MM-DD'
+    },
+    // below user set
     firstDayOfWeek: {
       default: 7,
       type: Number,
@@ -135,7 +146,7 @@ export default {
     const calendarMonth = now.getMonth()
     const firstYear = Math.floor(calendarYear / 10) * 10
     return {
-      panel: 'DATE',
+      panel: 'NONE',
       dates: [],
       calendarMonth,
       calendarYear,
@@ -153,11 +164,28 @@ export default {
         this.calendarMonth = now.getMonth()
       }
     },
+    timeType () {
+      const h = /h+/.test(this.$parent.format) ? '12' : '24'
+      const a = /A/.test(this.$parent.format) ? 'A' : 'a'
+      return [h, a]
+    },
     timeHeader () {
-      return this.value && new Date(this.value).toLocaleDateString()
+      if (this.type === 'time') {
+        return this.$parent.format
+      }
+      return this.value && formatDate(this.value, this.dateFormat)
+    },
+    yearHeader () {
+      return this.firstYear + ' ~ ' + (this.firstYear + 10)
     },
     months () {
       return this.t('months')
+    },
+    notBeforeTime () {
+      return this.getCriticalTime(this.notBefore)
+    },
+    notAfterTime () {
+      return this.getCriticalTime(this.notAfter)
     }
   },
   watch: {
@@ -170,56 +198,95 @@ export default {
       handler: 'init'
     },
     panel: {
-      immediate: true,
       handler: 'handelPanelChange'
     }
   },
   methods: {
-    handelPanelChange (panel) {
+    handelPanelChange (panel, oldPanel) {
+      this.$parent.$emit('panel-change', panel, oldPanel)
       if (panel === 'YEAR') {
         this.firstYear = Math.floor(this.calendarYear / 10) * 10
       } else if (panel === 'TIME') {
         this.$nextTick(() => {
-          [...this.$el.querySelectorAll('.mx-panel-time .mx-time-list')].forEach(el => {
+          const list = this.$el.querySelectorAll('.mx-panel-time .mx-time-list')
+          for (let i = 0, len = list.length; i < len; i++) {
+            const el = list[i]
             scrollIntoView(el, el.querySelector('.actived'))
-          })
+          }
         })
       }
     },
-    init () {
-      this.panel = 'DATE'
+    init (val) {
+      if (val) {
+        const type = this.type
+        if (type === 'month') {
+          this.showPanelMonth()
+        } else if (type === 'year') {
+          this.showPanelYear()
+        } else if (type === 'time') {
+          this.showPanelTime()
+        } else {
+          this.showPanelDate()
+        }
+      } else {
+        this.showPanelNone()
+      }
       this.updateNow(this.value)
     },
     // 根据value更新日历
     updateNow (value) {
       this.now = value ? new Date(value) : new Date()
     },
+    getCriticalTime (value) {
+      if (!value) {
+        return null
+      }
+      const date = new Date(value)
+      if (this.type === 'year') {
+        return new Date(date.getFullYear(), 0).getTime()
+      } else if (this.type === 'month') {
+        return new Date(date.getFullYear(), date.getMonth()).getTime()
+      } else if (this.type === 'date') {
+        return date.setHours(0, 0, 0, 0)
+      }
+      return date.getTime()
+    },
+    inBefore (time, startAt) {
+      startAt = startAt || this.startAt
+      return (this.notBeforeTime && time < this.notBeforeTime) ||
+        (startAt && time < this.getCriticalTime(startAt))
+    },
+    inAfter (time, endAt) {
+      endAt = endAt || this.endAt
+      return (this.notAfterTime && time > this.notAfterTime) ||
+        (endAt && time > this.getCriticalTime(endAt))
+    },
+    inDisabledDays (time) {
+      if (Array.isArray(this.disabledDays)) {
+        return this.disabledDays.some(v => this.getCriticalTime(v) === time)
+      } else if (typeof this.disabledDays === 'function') {
+        return this.disabledDays(new Date(time))
+      }
+      return false
+    },
+    isDisabledYear (year) {
+      const time = new Date(year, 0).getTime()
+      const maxTime = new Date(year + 1, 0).getTime() - 1
+      return this.inBefore(maxTime) || this.inAfter(time) || (this.type === 'year' && this.inDisabledDays(time))
+    },
+    isDisabledMonth (month) {
+      const time = new Date(this.calendarYear, month).getTime()
+      const maxTime = new Date(this.calendarYear, month + 1).getTime() - 1
+      return this.inBefore(maxTime) || this.inAfter(time) || (this.type === 'month' && this.inDisabledDays(time))
+    },
+    isDisabledDate (date) {
+      const time = new Date(date).getTime()
+      const maxTime = new Date(date).setHours(23, 59, 59, 999)
+      return this.inBefore(maxTime) || this.inAfter(time) || this.inDisabledDays(time)
+    },
     isDisabledTime (date, startAt, endAt) {
       const time = new Date(date).getTime()
-      const notBefore = this.notBefore && (time < new Date(this.notBefore))
-      const notAfter = this.notAfter && (time > new Date(this.notAfter))
-      startAt = startAt === undefined ? this.startAt : startAt
-      startAt = startAt && (time < new Date(startAt))
-      endAt = endAt === undefined ? this.endAt : endAt
-      endAt = endAt && (time > new Date(endAt))
-      return notBefore || notAfter || startAt || endAt
-    },
-    isDisabledDate (date, startAt, endAt) {
-      const time = new Date(date).getTime()
-      const notBefore = this.notBefore && (time < new Date(this.notBefore).setHours(0, 0, 0, 0))
-      const notAfter = this.notAfter && (time > new Date(this.notAfter).setHours(0, 0, 0, 0))
-      startAt = startAt === undefined ? this.startAt : startAt
-      startAt = startAt && (time < new Date(startAt).setHours(0, 0, 0, 0))
-      endAt = endAt === undefined ? this.endAt : endAt
-      endAt = endAt && (time > new Date(endAt).setHours(0, 0, 0, 0))
-      let disabledDays = false
-      if (Array.isArray(this.disabledDays)) {
-        disabledDays = this.disabledDays.some(v => new Date(v).setHours(0, 0, 0, 0) === time)
-      } else if (typeof this.disabledDays === 'function') {
-        disabledDays = this.disabledDays(new Date(date))
-      }
-
-      return notBefore || notAfter || disabledDays || startAt || endAt
+      return this.inBefore(time, startAt) || this.inAfter(time, endAt) || this.inDisabledDays(time)
     },
     selectDate (date) {
       if (this.type === 'datetime') {
@@ -240,22 +307,31 @@ export default {
             time = new Date(this.startAt)
           }
         }
-        this.$emit('select-time', time)
-        this.panel = 'TIME'
+        this.selectTime(time)
+        this.showPanelTime()
         return
       }
       this.$emit('select-date', date)
     },
     selectYear (year) {
       this.changeCalendarYear(year)
+      if (this.type.toLowerCase() === 'year') {
+        return this.selectDate(new Date(this.now))
+      }
       this.showPanelMonth()
     },
     selectMonth (month) {
       this.changeCalendarMonth(month)
+      if (this.type.toLowerCase() === 'month') {
+        return this.selectDate(new Date(this.now))
+      }
       this.showPanelDate()
     },
     selectTime (time) {
-      this.$emit('select-time', time)
+      this.$emit('select-time', time, false)
+    },
+    pickTime (time) {
+      this.$emit('select-time', time, true)
     },
     changeCalendarYear (year) {
       this.now = new Date(year, this.calendarMonth)
@@ -263,32 +339,56 @@ export default {
     changeCalendarMonth (month) {
       this.now = new Date(this.calendarYear, month)
     },
+    getSibling () {
+      const calendars = this.$parent.$children.filter(v => v.$options.name === this.$options.name)
+      const index = calendars.indexOf(this)
+      const sibling = calendars[index ^ 1]
+      return sibling
+    },
     handleIconMonth (flag) {
-      this.changeCalendarMonth(this.calendarMonth + flag)
+      const month = this.calendarMonth
+      this.changeCalendarMonth(month + flag)
+      this.$parent.$emit('change-calendar-month', {
+        month,
+        flag,
+        vm: this,
+        sibling: this.getSibling()
+      })
     },
     handleIconYear (flag) {
       if (this.panel === 'YEAR') {
         this.changePanelYears(flag)
       } else {
-        this.changeCalendarYear(this.calendarYear + flag)
+        const year = this.calendarYear
+        this.changeCalendarYear(year + flag)
+        this.$parent.$emit('change-calendar-year', {
+          year,
+          flag,
+          vm: this,
+          sibling: this.getSibling()
+        })
       }
     },
     handleBtnYear () {
-      if (this.panel === 'YEAR') {
-        this.showPanelDate()
-      } else {
-        this.showPanelYear()
-      }
+      this.showPanelYear()
     },
     handleBtnMonth () {
-      if (this.panel === 'MONTH') {
-        this.showPanelDate()
-      } else {
-        this.showPanelMonth()
+      this.showPanelMonth()
+    },
+    handleTimeHeader () {
+      if (this.type === 'time') {
+        return
       }
+      this.showPanelDate()
     },
     changePanelYears (flag) {
       this.firstYear = this.firstYear + flag * 10
+    },
+    showPanelNone () {
+      this.panel = 'NONE'
+    },
+    showPanelTime () {
+      this.panel = 'TIME'
     },
     showPanelDate () {
       this.panel = 'DATE'