+++ src/index.vue
@@ -12,9 +12,10 @@
     <div class="mx-input-wrapper"
       @click="showPopup">
       <input
-        :class="inputClass"
         ref="input"
         type="text"
+        autocomplete="off"
+        :class="inputClass"
         :name="inputName"
         :disabled="disabled"
         :readonly="!editable"
@@ -48,11 +49,11 @@
       ref="calendar">
       <slot name="header">
         <div class="mx-shortcuts-wrapper"
-          v-if="range && innnerShortcuts.length">
+          v-if="range && innerShortcuts.length">
           <button
             type="button"
             class="mx-shortcuts"
-            v-for="(range, index) in innnerShortcuts"
+            v-for="(range, index) in innerShortcuts"
             :key="index"
             @click="selectRange(range)">{{range.text}}</button>
         </div>
@@ -60,6 +61,8 @@
       <calendar-panel
         v-if="!range"
         v-bind="$attrs"
+        :type="innerType"
+        :date-format="innerDateFormat"
         :value="currentValue"
         :visible="popupVisible"
         @select-date="selectDate"
@@ -69,6 +72,8 @@
         <calendar-panel
           style="box-shadow:1px 0 rgba(0, 0, 0, .1)"
           v-bind="$attrs"
+          :type="innerType"
+          :date-format="innerDateFormat"
           :value="currentValue[0]"
           :end-at="currentValue[1]"
           :start-at="null"
@@ -77,6 +82,8 @@
           @select-time="selectStartTime"></calendar-panel>
         <calendar-panel
           v-bind="$attrs"
+          :type="innerType"
+          :date-format="innerDateFormat"
           :value="currentValue[1]"
           :start-at="currentValue[0]"
           :end-at="null"
@@ -99,7 +106,7 @@
 <script>
 import fecha from 'fecha'
 import clickoutside from '@/directives/clickoutside'
-import { isValidDate, isValidRange, isDateObejct, isPlainObject } from '@/utils/index'
+import { isValidDate, isValidRange, isDateObejct, isPlainObject, formatDate, parseDate } from '@/utils/index'
 import CalendarPanel from './calendar.vue'
 import locale from '@/mixins/locale'
 import Languages from '@/locale/languages'
@@ -125,6 +132,13 @@ export default {
     format: {
       type: String,
       default: 'YYYY-MM-DD'
+    },
+    dateFormat: {
+      type: String // format the time header and date tooltip
+    },
+    type: {
+      type: String,
+      default: 'date' // ['date', 'datetime'] zendy added 'month', 'year', mxie added "time"
     },
     range: {
       type: Boolean,
@@ -225,7 +239,10 @@ export default {
     showClearIcon () {
       return !this.disabled && this.clearable && (this.range ? isValidRange(this.value) : isValidDate(this.value))
     },
-    innnerShortcuts () {
+    innerType () {
+      return String(this.type).toLowerCase()
+    },
+    innerShortcuts () {
       if (Array.isArray(this.shortcuts)) {
         return this.shortcuts
       }
@@ -264,6 +281,15 @@ export default {
         }
       ]
       return arr
+    },
+    innerDateFormat () {
+      if (this.dateFormat) {
+        return this.dateFormat
+      }
+      if (this.innerType === 'date') {
+        return this.format
+      }
+      return this.format.replace(/[Hh]+.*[msSaAZ]|\[.*?\]/g, '').trim() || 'YYYY-MM-DD'
     }
   },
   methods: {
@@ -272,20 +298,10 @@ export default {
       this.displayPopup()
     },
     stringify (date, format) {
-      try {
-        format = format || this.format
-        return fecha.format(new Date(date), format)
-      } catch (e) {
-        return ''
-      }
+      return formatDate(date, format || this.format)
     },
     parseDate (value, format) {
-      try {
-        format = format || this.format
-        return fecha.parse(value, format)
-      } catch (e) {
-        return false
-      }
+      return parseDate(value, format || this.format)
     },
     dateEqual (a, b) {
       return isDateObejct(a) && isDateObejct(b) && a.getTime() === b.getTime()
@@ -304,6 +320,7 @@ export default {
       const date = this.range ? [null, null] : null
       this.currentValue = date
       this.updateDate(true)
+      this.$emit('clear')
     },
     confirmDate () {
       const valid = this.range ? isValidRange(this.currentValue) : isValidDate(this.currentValue)
@@ -348,9 +365,9 @@ export default {
         this.updateDate()
       }
     },
-    selectTime (time) {
+    selectTime (time, close) {
       this.currentValue = time
-      this.updateDate()
+      this.updateDate() && close && this.closePopup()
     },
     selectStartTime (time) {
       this.selectStartDate(time)
@@ -401,7 +418,7 @@ export default {
       const value = event.target.value
       if (this.editable && this.userInput !== null) {
         const calendar = this.$children[0]
-        const checkDate = calendar.type === 'date' ? calendar.isDisabledDate : calendar.isDisabledTime
+        const checkDate = calendar.isDisabledTime
         if (this.range) {
           const range = value.split(` ${this.rangeSeparator} `)
           if (range.length === 2) {